Create helper.js
This was first used in the redesign of the Bluedot website. It was used as a means to dynamically change the active menu class (among other things) in the main navigation menu or element. It was done so I wouldn't have different versions of the main-navigation.php file and it's corresponding css file. It was used firstly in casestudies.php and then in contact.php.