Background/Context to the code
The file or logic built or constructed to be completely dynamic. Previously it had hardcoded values like const casestudies = document.querySelector('[href="casestudies.php"]'); which has been replaced with const casestudies = document.querySelector("[href=" + CSS.escape(lastsegment) + "]");. It wouldn't require you to go into the code and change the values manually because the values change in relation to the url.